requirements.txt
streamlit>=1.21.0
altair<5
pytest
google-genai>=0.8.0
python-dotenv>=1.0.0
